+++ packages/address/src/Address.ts
@@ -6,6 +6,8 @@ export interface Address<out V> extends ReadonlyAddress<V> {
 
   filter(predicate: BinaryPredicate<V, void>): Address<V>;
 
+  filter<W extends V>(predicate: BinaryPredicate<W, void>): Address<W>;
+
   map<W>(mapping: Mapping<V, W>): Address<W>;
 
   remove(value: V): Address<V>;