+++ packages/sequence/src/MutableSequence.ts
@@ -41,6 +41,8 @@ export class MutableSequence<out V> extends ASequence<V> {
     return MutableSequence.ofArray([...this.sequence]);
   }
 
+  public filter(predicate: BinaryPredicate<V, number>): MutableSequence<V>;
+  public filter<W extends V>(predicate: BinaryPredicate<W, number>): MutableSequence<W>;
   public filter(predicate: BinaryPredicate<V, number>): MutableSequence<V> {
     return MutableSequence.ofArray(this.filterInternal(predicate));
   }