+++ packages/sequence/src/Sequence.ts
@@ -6,6 +6,8 @@ export interface Sequence<out V> extends ReadonlySequence<V> {
 
   filter(predicate: BinaryPredicate<V, number>): Sequence<V>;
 
+  filter<W extends V>(predicate: BinaryPredicate<W, number>): Sequence<W>;
+
   map<W>(mapping: Mapping<V, W>): Sequence<W>;
 
   remove(key: number): Sequence<V>;